requirements.txt
#fastapi==0.103.1
#uvicorn==0.22.0
#opencv-python
#pytesseract
#deepface
#openai
#numpy
#pillow
#pydantic
#tf-keras
google-cloud-vision
#httpx
#google.cloud
#azure-cognitiveservices-vision-face
#boto3
#unidecode
#uuid
#face_recognition
#mtcnn